# packages/backend/src/api/openapi.yaml
openapi: 3.0.0
info:
  title: Scroblarr API
  version: 1.0.0
  description: |
    API documentation for Scroblarr - Media scrobbling service for Plex, Jellyfin, and Emby.

    This API allows you to manage users, sync media watching history across services, and configure integrations with TVTime and Trakt.
servers:
  - url: http://localhost:3000
    description: Development server
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token (Plex or Jellyfin access token)
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Global API key for API authentication
  schemas:
...
    Success:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          nullable: true
...
    TVTimeStatus:
      type: object
      properties:
        linked:
          type: boolean
        email:
          type: string
          nullable: true
        username:
          type: string
          nullable: true
    TVTimeProfile:
      type: object
      properties:
        id:
          type: string
          nullable: true
        username:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        image:
          type: string
          nullable: true
    TraktStatus:
      type: object
      properties:
        linked:
          type: boolean
        username:
          type: string
          nullable: true
        image:
          type: string
          nullable: true
        hasCredentials:
          type: boolean
...
paths:
...
  /api/v1/tvtime/link:
    post:
      summary: Link TVTime account
      tags: [TVTime]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Account linked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /api/v1/tvtime/unlink:
    post:
      summary: Unlink TVTime account
      tags: [TVTime]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Account unlinked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /api/v1/tvtime/status:
    get:
      summary: Get TVTime account status
      tags: [TVTime]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Account status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TVTimeStatus'
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /api/v1/tvtime/profile:
    get:
      summary: Get TVTime profile
      tags: [TVTime]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TVTimeProfile'
        '400':
          description: Account not linked
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /api/v1/trakt/authorize:
    get:
      summary: Get Trakt authorization URL
      tags: [Trakt]
      security:
        - bearerAuth: []
      parameters:
        - name: clientId
          in: query
          schema:
            type: string
        - name: clientSecret
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Authorization URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  authUrl:
                    type: string
                    format: uri
        '400':
          description: Bad request (missing credentials)
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /api/v1/trakt/link:
    post:
      summary: Link Trakt account
      tags: [Trakt]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
              properties:
                code:
                  type: string
                clientId:
                  type: string
                clientSecret:
                  type: string
      responses:
        '200':
          description: Account linked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          description: Validation error or missing credentials
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /api/v1/trakt/unlink:
    post:
      summary: Unlink Trakt account
      tags: [Trakt]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Account unlinked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /api/v1/trakt/status:
    get:
      summary: Get Trakt account status
      tags: [Trakt]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Account status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraktStatus'
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
...